backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
bcrypt
python-multipart
email-validator
